Factor out address_form_fields.py.
